Added bluetooth and encoder libraries

+++ src/bmx_utilities.c
@@ -7,6 +7,6 @@ void float_to_2ints(float input, int *output, unsigned int dec)
   int dec_places = 1;
   for(int i = 0; i < dec; i++) {dec_places *= 10;}
 
-  output[0] = (int) input; // number
+  output[0] = input; // number
   output[1] = abs(input*dec_places - output[0]*dec_places); // decimal
 }
